+++ frontend/src/component/pages/shared/UserShortLinksSection.tsx
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import { PageControl } from '../../ui/PageControl';
 import { Table } from '../../ui/Table';
-import { Url } from '../../../entity/Url';
+import { ShortLink } from '../../../entity/ShortLink';
 import { IPagedShortLinks } from '../../../service/ShortLink.service';
 import './UserShortLinksSection.scss';
 import { Section } from '../../ui/Section';
@@ -53,7 +53,7 @@ export class UserShortLinksSection extends Component<IProps> {
 
   private createTableRows = () => {
     const { shortLinks } = this.props.pagedShortLinks!;
-    return shortLinks.map((shortLink: Url) => {
+    return shortLinks.map((shortLink: ShortLink) => {
       return [
         this.renderLongLink(shortLink.originalUrl),
         this.renderAlias(shortLink.alias)
