+++ frontend/src/entity/ShortLink.tsx
@@ -1,4 +1,4 @@
-export interface Url {
+export interface ShortLink {
   originalUrl: string;
   alias: string;
 }
