+++ frontend/src/service/Search.service.ts
@@ -1,15 +1,15 @@
-import { Url } from '../entity/Url';
+import { ShortLink } from '../entity/ShortLink';
 import { Err } from './Error.service';
 
 export class SearchService {
-  getAutoCompleteSuggestions(alias: String): Promise<Array<Url>> {
+  getAutoCompleteSuggestions(alias: String): Promise<Array<ShortLink>> {
     return new Promise(async (resolve, reject) => {
       resolve(await this.invokeSearchShortLinkApi(alias));
     });
   }
 
-  private async invokeSearchShortLinkApi(alias: String): Promise<Array<Url>> {
-    return new Promise<Array<Url>>(
+  private async invokeSearchShortLinkApi(alias: String): Promise<Array<ShortLink>> {
+    return new Promise<Array<ShortLink>>(
       (resolve, reject: (errCodes: Err[]) => any) => {
         if (alias === '') {
           resolve([]);
